# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/auth/login:
    post:
      operationId: dashboardLogin
      tags: [dashboard-auth]
      summary: Login to dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard/products:
    get:
      operationId: listDashboardProducts
      tags: [dashboard-products]
      summary: List and search products with filters
      parameters:
        - name: q
          in: query
          schema:
            type: ["string", "null"]
          description: Search by title, handle, SKU, or GTIN
        - name: market
          in: query
          schema:
            type: ["string", "null"]
        - name: language
          in: query
          schema:
            type: ["string", "null"]
        - name: brand
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: availability
          in: query
          schema:
            type: ["string", "null"]
        - name: lifecycle
          in: query
          schema:
            type: ["string", "null"]
          description: Lifecycle label (outlet, discontinued, bestseller, exhibition_model)
        - name: eligible
          in: query
          schema:
            type: ["boolean", "null"]
        - name: minQuality
          in: query
          schema:
            type: ["number", "null"]
        - name: maxQuality
          in: query
          schema:
            type: ["number", "null"]
        - name: limit
          in: query
          schema:
            type: ["number", "null"]
        - name: offset
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
  /dashboard/products/{id}:
    get:
      operationId: getDashboardProduct
      tags: [dashboard-products]
      summary: Full product debug view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product UUID
      responses:
        "200":
          description: Product debug data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDebug"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Google ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginCredentials:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AuthSession:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated

    # ── Overview ───────────────────────────────────────────────────────────────
...
    ProductSummary:
      type: object
      required:
        - id
        - handle
        - vendor
        - variantCount
        - eligibleVariantCount
        - availability
        - avgQualityScore
        - markets
        - updatedAt
      properties:
        id:
          type: string
        handle:
          type: string
        title:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
        productType:
          type: ["string", "null"]
        variantCount:
          type: number
        eligibleVariantCount:
          type: number
        availability:
          type: string
        avgQualityScore:
          type: ["number", "null"]
        markets:
          type: array
          items:
            type: string
        lifecycleFlags:
          type: array
          items:
            type: string
        updatedAt:
          type: string
    ProductList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummary"
        total:
          type: number
    ProductDebug:
      type: object
      required:
        - product
        - variants
        - translations
        - marketVariants
        - images
        - inventory
        - feedItems
        - recommendations
      properties:
        product:
          $ref: "#/components/schemas/ProductSource"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantDetail"
        translations:
          type: array
          items:
            $ref: "#/components/schemas/ProductTranslation"
        marketVariants:
          type: array
          items:
            $ref: "#/components/schemas/MarketVariantDetail"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ImageDetail"
        inventory:
          type: array
          items:
            $ref: "#/components/schemas/InventoryLevel"
        feedItems:
          type: array
          items:
            $ref: "#/components/schemas/FeedItemDetail"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationDetail"
    ProductSource:
      type: object
      required: [id, shopifyGid, handle, status]
      properties:
        id:
          type: string
        shopifyGid:
          type: string
        shopifyId:
          type: ["string", "null"]
        handle:
          type: string
        vendor:
          type: ["string", "null"]
        productType:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        publishedAt:
          type: ["string", "null"]
        sourceUpdatedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductTranslation:
      type: object
      required: [language, title]
      properties:
        language:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        handle:
          type: ["string", "null"]
    VariantDetail:
      type: object
      required: [id, shopifyGid, title, sku]
      properties:
        id:
          type: string
        shopifyGid:
          type: string
        shopifyId:
          type: ["string", "null"]
        title:
          type: string
        sku:
          type: ["string", "null"]
        gtin:
          type: ["string", "null"]
        mpn:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        weightUnit:
          type: string
        available:
          type: boolean
        googleCategory:
          type: ["string", "null"]
        metaCategory:
          type: ["string", "null"]
        shippingClass:
          type: ["string", "null"]
        returnClass:
          type: ["string", "null"]
        outlet:
          type: ["boolean", "null"]
        discontinued:
          type: ["boolean", "null"]
        bestseller:
          type: ["boolean", "null"]
        exhibitionModel:
          type: ["boolean", "null"]
        material:
          type: array
          items:
            type: string
        style:
          type: array
          items:
            type: string
        room:
          type: array
          items:
            type: string
    MarketVariantDetail:
      type: object
      required: [variantId, marketCode, availability, isEligible]
      properties:
        variantId:
          type: string
        marketCode:
          type: string
        priceAmount:
          type: ["number", "null"]
        compareAtPriceAmount:
          type: ["number", "null"]
        priceCurrency:
          type: ["string", "null"]
        availability:
          type: string
        productUrl:
          type: ["string", "null"]
        isEligible:
          type: boolean
        exclusionReason:
          type: ["string", "null"]
    ImageDetail:
      type: object
      required: [id, url, position]
      properties:
        id:
          type: string
        url:
          type: string
        altText:
          type: ["string", "null"]
        position:
          type: number
        imageType:
          type: ["string", "null"]
        width:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        whiteBgScore:
          type: ["number", "null"]
        solidBgScore:
          type: ["number", "null"]
        alphaRatio:
          type: ["number", "null"]
        edgeDensity:
          type: ["number", "null"]
        resolutionScore:
          type: ["number", "null"]
        isClassified:
          type: boolean
        classifiedAt:
          type: ["string", "null"]
    InventoryLevel:
      type: object
      required: [shopifyLocationId, locationName, available]
      properties:
        shopifyLocationId:
          type: string
        locationName:
          type: ["string", "null"]
        available:
          type: number
    FeedItemDetail:
      type: object
      required: [variantId, marketCode, language, channel, isEligible]
      properties:
        variantId:
          type: string
        marketCode:
          type: string
        language:
          type: string
        channel:
          type: string
        isEligible:
          type: boolean
        exclusionReason:
          type: ["string", "null"]
        dataQualityScore:
          type: ["number", "null"]
        generatedAt:
          type: ["string", "null"]
    RecommendationDetail:
      type: object
      required: [marketCode, relatedProductIds, complementaryProductIds]
      properties:
        marketCode:
          type: string
        relatedProductIds:
          type: array
          items:
            type: string
        complementaryProductIds:
          type: array
          items:
            type: string
        generatedAt:
          type: ["string", "null"]

    # ── Google ─────────────────────────────────────────────────────────────────
...