# lib/api-zod/src/generated/types/productSource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductSource {
  id: string;
  shopifyGid: string;
  /** @nullable */
  shopifyId?: string | null;
  handle: string;
  /** @nullable */
  vendor?: string | null;
  /** @nullable */
  productType?: string | null;
  tags?: string[];
  status: string;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  sourceUpdatedAt?: string | null;
  createdAt?: string;
  updatedAt?: string;
}